test: split context menu items tests into separate files
Move interaction tests to items-interactions.test.js and rendering tests
to items-rendering.test.js, which also takes over items-theme.test.js.
Group the remaining tests into nested suites, and give each file only the
items that its own tests need.

The shared beforeEach opened both the menu and a sub-menu, which most
tests had to undo first. Each suite now opens what it needs, using a new
openSubMenu() helper.

Use sendKeys in the interaction tests, so that they also cover the
browser behavior the component relies on, such as moving focus.

Also restore the style element and the scroll position in the scrolling
suite, which leaked into the following tests.
